Use an interface for SyncService and skip doing fork choice while we are syncing so we can focus resources on catching up.
